+++ scala3/src/test/scala/testing/TestWatch.scala
@@ -9,12 +9,19 @@ object TestWatch:
   private var started: Boolean = false
   
   def start(): Unit = synchronized {
-    if (!started) then
-      t0 = Instant.now
-      started = true
+    if (!started) then restart()
+  }
+  
+  def restart(): Unit = synchronized {
+    t0 = Instant.now
+    started = true
   }
 
   def dt: Duration = if started then Duration.between(t0, Instant.now) else Duration.ZERO
+  
+  def timePassedSec: String =
+    val diff = dt
+    String.format("%02d.%03d", diff.toSeconds, diff.toMillisPart)
   
   def timePassed: String =
     val diff = dt
